+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Allow user-defined `profiles` location (#27)
 
+### Added
+
+* Available as `conda` package via `conda-forge` (#34)
+
+    ```console
+    $ conda install -c conda-forge radiant-mlhub
+    ```
+
 ## \[v0.1.2\]
 
 ### Fixed