+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## \[Unreleased\]
+
+### Fixed
+
+* Allow user-defined `profiles` location (#27)
+
+### Added
+
+* Available as `conda` package via `conda-forge` (#34)
+
+    ```console
+    $ conda install -c conda-forge radiant-mlhub
+    ```
+
 ## \[v0.1.2\]
 
 ### Fixed