+++ CHANGELOG.md
@@ -18,6 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     $ conda install -c conda-forge radiant-mlhub
     ```
 
+## \[v0.1.3\]
+
+### Fixed
+
+* New attributes in the `/dataset` response will no longer break the `Dataset` class
+  ([#42](https://github.com/radiantearth/radiant-mlhub/pull/42))
+
 ## \[v0.1.2\]
 
 ### Fixed